+++ packages/flutter/lib/rendering.dart
@@ -63,7 +63,6 @@ export 'src/rendering/table.dart';
 export 'src/rendering/table_border.dart';
 export 'src/rendering/texture.dart';
 export 'src/rendering/tweens.dart';
-export 'src/rendering/unconstrained_box.dart';
 export 'src/rendering/view.dart';
 export 'src/rendering/viewport.dart';
 export 'src/rendering/viewport_offset.dart';
